+++ src/main/java/roomescape/domain/Reservation.java
@@ -1,4 +1,4 @@
-package roomescape.reservation;
+package roomescape.domain;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
